[geombong/issue-tracker#99] Refactor: Change Package Location
## Description
- 클래스가 위치 변경 및 새로운 request, response 패키지 생성
