+++ BE/src/test/java/team20/issuetracker/domain/milestone/MilestoneCreateTest.java
@@ -6,7 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
-import team20.issuetracker.domain.milestone.requestDto.SaveMilestoneDto;
+import team20.issuetracker.domain.milestone.request.SaveMilestoneDto;
 import team20.issuetracker.service.MilestoneService;
 
 import java.time.LocalDateTime;
